azfunction/etlfunctionNew/requirements.txt
azure-functions
azure-storage-blob
Click
gopro2gpx==0.1
moviepy
shapely
pandas
psycopg2-binary
pyproj
gpxpy
requests
tqdm